Nest files roughly based on their dependency tree
