+++ src/main/WindowManager/MainWindow/RpcRequestIpcEvent/BcoreRpcConfigurations/Cookie.ts
