+++ src/main/WindowManager/MainWindow/RpcRequestIpcEvent/MainRpcClient/MainRpcClient.ts
@@ -1,0 +1,76 @@
+import http from "http";
+import { RPC_ERROR } from "_c/constants";
+import { RpcRequest } from "_t/RpcRequests";
+import { RawRpcResponse } from "_t/RpcResponses";
+import { ExtractedRpcResponse } from "_t/typeHelpers";
+import { ErrorWithCode } from "_c/ErrorWithCode";
+
+type RequestParams = {
+  url: string;
+  options: http.RequestOptions;
+  body: any;
+};
+
+class MainRpcClient {
+  private httpRequest = (
+    requestParams: RequestParams,
+  ): Promise<http.IncomingMessage & { data: string }> => {
+    const { url, options, body } = requestParams;
+
+    return new Promise((resolve, reject) => {
+      const nodeRequest = http.request(url, options, response => {
+        response.setEncoding("utf8");
+
+        let data = "";
+
+        response.on("data", dataChunk => {
+          data += dataChunk;
+        });
+
+        response.on("end", () => {
+          resolve(Object.assign(response, { data }));
+        });
+      });
+
+      nodeRequest.write(JSON.stringify(body));
+
+      nodeRequest.on("error", reject);
+
+      nodeRequest.end();
+    });
+  };
+
+  call = async <TRpcRequest extends RpcRequest>(
+    rpcRequest: TRpcRequest,
+    rpcConfigurations: {
+      username: string;
+      password: string;
+      serverUrl: string;
+    },
+  ): Promise<ExtractedRpcResponse<TRpcRequest>> => {
+    const { method, params = [] } = rpcRequest;
+    const { username, password, serverUrl } = rpcConfigurations;
+
+    const options = {
+      method: "POST",
+      auth: `${username}:${password}`,
+      headers: { "content-type": "text/plain" },
+    };
+    const body = { jsonrpc: "1.0", id: "N/A", method, params };
+
+    const response = await this.httpRequest({ url: serverUrl, options, body });
+
+    if (response.statusCode === 401 || response.statusCode === 403) {
+      throw new ErrorWithCode(
+        "RPC server unauthorized request",
+        RPC_ERROR.unauthorized,
+      );
+    }
+
+    const { result, error }: RawRpcResponse = JSON.parse(response.data);
+
+    return { result, error, method } as ExtractedRpcResponse<TRpcRequest>;
+  };
+}
+
+export const mainRpcClient = new MainRpcClient();
