+++ src/main/WindowManager/MainWindow/RpcRequestIpcEvent/RpcRequestIpcEvent.ts
@@ -1,9 +1,9 @@
 import { BITCOIN_CORE_RPC_ERROR, NODE_ERROR, RPC_ERROR } from "_c/constants";
 import { featureFlags } from "_f/featureFlags";
 import { btcdRpcConfigurations } from "_m/common/BtcdRpcConfigurations";
-import { mainRpcClient } from "_m/mainRpcClient/mainRpcClient";
-import { bcoreRpcConfigurations } from "_m/MainWindow/RpcRequestIpcEvent/BcoreRpcConfigurations/BcoreRpcConfigurations";
-import { windowManager } from "_m/WindowManager";
+import { mainRpcClient } from "_m/WindowManager/MainWindow/RpcRequestIpcEvent/MainRpcClient/MainRpcClient";
+import { bcoreRpcConfigurations } from "_m/WindowManager/MainWindow/RpcRequestIpcEvent/BcoreRpcConfigurations/BcoreRpcConfigurations";
+import { windowManager } from "_m/WindowManager/WindowManager";
 import { SendableMessageToMain } from "_t/IpcMessages";
 import { RpcResponse } from "_t/RpcResponses";
 import { PromiseType } from "_t/typeHelpers";
@@ -46,7 +46,7 @@ export class RpcRequestIpcEvent {
         rpcConfigurations = btcdRpcConfigurations;
       }
 
-      response = await mainRpcClient(data.payload, rpcConfigurations);
+      response = await mainRpcClient.call(data.payload, rpcConfigurations);
     } catch (error) {
       const errorResponse = {
         result: null,
