+++ src/main/common/ErrorDialog.ts
