+++ src/main/common/Utils.ts
