--- src/main/mainRpcClient/isRpcMethodAllowed.ts
@@ -1,14 +1,0 @@
-import { RpcRequest } from "_t/RpcRequests";
-
-export const isRpcMethodAllowed = (methodName: RpcRequest["method"]) => {
-  const allowedMethods: RpcRequest["method"][] = [
-    "getblockchaininfo",
-    "getblock",
-    "getblockhash",
-    "getrawtransaction",
-    "getmempoolinfo",
-    "uptime",
-  ];
-
-  return allowedMethods.includes(methodName);
-};
